PlateManagerTest
